added a GameSection file and all is working except setup name enter still looks ugly
